Move utils classes to Shaarli\Helper namespace and folder
