+++ tests/helper/FileUtilsTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Shaarli;
+namespace Shaarli\Helper;
 
 use Exception;
 use Shaarli\Exceptions\IOException;
